add RenderAvatar component
